Fix some bugs, start implementing WAHBitSetIterator
git-svn-id: https://guust.tuxes.nl/svn/bas/studie/0910-afstuderen/Implementations/c++@413 840f488b-753d-4fb4-b916-1b28b8a18e6e

+++ WAHStackTC/src/WAHStackTC.h
@@ -37,7 +37,7 @@ public:
 
 	void computeTransitiveClosure();
 	string tcToString();
-	int countNumberOfEdgesInTC();
+	long countNumberOfEdgesInTC();
 };
 
 #endif /* WAHSTACKTC_H_ */
